demo/usecase-2/passenger-info/seat-info-proxy/requirements.txt
psycopg2>=2.9.1,<3.0
SQLAlchemy>=1.4.18,<2.0
numpy>=1.20.3,<2.0
PyYAML>=5.4.1,<6.0
edgefarm-application>=0.2.0,<1.0
schedule>=1.1.0,<2.0
fastapi-asyncpg>=1.0.1,<2.0
